requirements.txt
# Core dependencies
pyserial>=3.5
Pillow>=10.0.0
pyyaml>=6.0.1
pygments>=2.15.1

# Development dependencies
black>=23.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Documentation
typing-extensions>=4.5.0

# Testing
pytest-html>=4.0.0
pytest-xdist>=3.2.0
